Rename package to have consistent naming
