+++ angular.json
@@ -3,7 +3,7 @@
   "version": 1, 
   "newProjectRoot": "projects",
   "projects": {
-    "fabricjs-editor": {
+    "angular-editor-fabric-js-app": {
       "projectType": "application",
       "schematics": {
         "@schematics/angular:component": {
@@ -17,7 +17,7 @@
         "build": {
           "builder": "@angular-devkit/build-angular:browser",
           "options": {
-            "outputPath": "dist/fabricjs-editor",
+            "outputPath": "dist/angular-editor-fabric-js",
             "index": "src/index.html",
             "main": "src/main.ts",
             "polyfills": "src/polyfills.ts",
@@ -68,18 +68,18 @@
         "serve": {
           "builder": "@angular-devkit/build-angular:dev-server",
           "options": {
-            "browserTarget": "fabricjs-editor:build"
+            "browserTarget": "angular-editor-fabric-js:build"
           },
           "configurations": {
             "production": {
-              "browserTarget": "fabricjs-editor:build:production"
+              "browserTarget": "angular-editor-fabric-js:build:production"
             }
           }
         },
         "extract-i18n": {
           "builder": "@angular-devkit/build-angular:extract-i18n",
           "options": {
-            "browserTarget": "fabricjs-editor:build"
+            "browserTarget": "angular-editor-fabric-js:build"
           }
         },
         "test": {
@@ -116,48 +116,48 @@
           "builder": "@angular-devkit/build-angular:protractor",
           "options": {
             "protractorConfig": "e2e/protractor.conf.js",
-            "devServerTarget": "fabricjs-editor:serve"
+            "devServerTarget": "angular-editor-fabric-js:serve"
           },
           "configurations": {
             "production": {
-              "devServerTarget": "fabricjs-editor:serve:production"
+              "devServerTarget": "angular-editor-fabric-js:serve:production"
             }
           }
         }
       }
     },
-    "fabricjs-library": {
+    "angular-editor-fabric-js": {
       "projectType": "library",
-      "root": "projects/fabricjs-library",
-      "sourceRoot": "projects/fabricjs-library/src",
+      "root": "projects/angular-editor-fabric-js",
+      "sourceRoot": "projects/angular-editor-fabric-js/src",
       "prefix": "lib",
       "architect": {
         "build": {
           "builder": "@angular-devkit/build-ng-packagr:build",
           "options": {
-            "tsConfig": "projects/fabricjs-library/tsconfig.lib.json",
-            "project": "projects/fabricjs-library/ng-package.json"
+            "tsConfig": "projects/angular-editor-fabric-js/tsconfig.lib.json",
+            "project": "projects/angular-editor-fabric-js/ng-package.json"
           },
           "configurations": {
             "production": {
-              "tsConfig": "projects/fabricjs-library/tsconfig.lib.prod.json"
+              "tsConfig": "projects/angular-editor-fabric-js/tsconfig.lib.prod.json"
             }
           }
         },
         "test": {
           "builder": "@angular-devkit/build-angular:karma",
           "options": {
-            "main": "projects/fabricjs-library/src/test.ts",
-            "tsConfig": "projects/fabricjs-library/tsconfig.spec.json",
-            "karmaConfig": "projects/fabricjs-library/karma.conf.js"
+            "main": "projects/angular-editor-fabric-js/src/test.ts",
+            "tsConfig": "projects/angular-editor-fabric-js/tsconfig.spec.json",
+            "karmaConfig": "projects/angular-editor-fabric-js/karma.conf.js"
           }
         },
         "lint": {
           "builder": "@angular-devkit/build-angular:tslint",
           "options": {
             "tsConfig": [
-              "projects/fabricjs-library/tsconfig.lib.json",
-              "projects/fabricjs-library/tsconfig.spec.json"
+              "projects/angular-editor-fabric-js/tsconfig.lib.json",
+              "projects/angular-editor-fabric-js/tsconfig.spec.json"
             ],
             "exclude": [
               "**/node_modules/**"
@@ -166,5 +166,5 @@
         }
       }
     }},
-  "defaultProject": "fabricjs-editor"
+  "defaultProject": "angular-editor-fabric-js"
 }
