+++ projects/angular-editor-fabric-js/README.md
@@ -1,0 +1,24 @@
+# FabricjsLibrary
+
+This library was generated with [Angular CLI](https://github.com/angular/angular-cli) version 9.0.7.
+
+## Code scaffolding
+
+Run `ng generate component component-name --project angular-editor-fabric-js` to generate a new component. You can also use `ng generate directive|pipe|service|class|guard|interface|enum|module --project angular-editor-fabric-js`.
+> Note: Don't forget to add `--project angular-editor-fabric-js` or else it will be added to the default project in your `angular.json` file. 
+
+## Build
+
+Run `ng build angular-editor-fabric-js` to build the project. The build artifacts will be stored in the `dist/` directory.
+
+## Publishing
+
+After building your library with `ng build angular-editor-fabric-js`, go to the dist folder `cd dist/angular-editor-fabric-js` and run `npm publish`.
+
+## Running unit tests
+
+Run `ng test angular-editor-fabric-js` to execute the unit tests via [Karma](https://karma-runner.github.io).
+
+## Further help
+
+To get more help on the Angular CLI use `ng help` or go check out the [Angular CLI README](https://github.com/angular/angular-cli/blob/master/README.md).
