+++ projects/angular-editor-fabric-js/ng-package.json
@@ -1,6 +1,6 @@
 {
   "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
-  "dest": "../../dist/fabricjs-library",
+  "dest": "../../dist/angular-editor-fabric-js",
   "lib": {
     "entryFile": "src/public-api.ts"
   }
