+++ projects/angular-editor-fabric-js/src/lib/angular-editor-fabric-js.component.html
