+++ projects/angular-editor-fabric-js/src/lib/angular-editor-fabric-js.module.ts
@@ -3,7 +3,7 @@ import { BrowserModule } from '@angular/platform-browser';
 import { FormsModule } from '@angular/forms';
 import { ColorPickerModule } from 'ngx-color-picker';
 
-import { FabricjsLibraryComponent } from './fabricjs-library.component';
+import { FabricjsLibraryComponent } from './angular-editor-fabric-js.component';
 
 @NgModule({
   declarations: [FabricjsLibraryComponent],
