+++ projects/angular-editor-fabric-js/src/lib/angular-editor-fabric-js.service.ts
