+++ projects/angular-editor-fabric-js/src/public-api.ts
@@ -1,0 +1,7 @@
+/*
+ * Public API Surface of angular-editor-fabric-js
+ */
+
+export * from './lib/angular-editor-fabric-js.service';
+export * from './lib/angular-editor-fabric-js.component';
+export * from './lib/angular-editor-fabric-js.module';
