+++ projects/angular-editor-fabric-js/tsconfig.lib.json
