+++ projects/angular-editor-fabric-js/tsconfig.lib.prod.json
