+++ projects/angular-editor-fabric-js/tsconfig.spec.json
