+++ projects/angular-editor-fabric-js/tslint.json
