--- projects/fabricjs-library/src/public-api.ts
@@ -1,7 +1,0 @@
-/*
- * Public API Surface of fabricjs-library
- */
-
-export * from './lib/fabricjs-library.service';
-export * from './lib/fabricjs-library.component';
-export * from './lib/fabricjs-library.module';
