+++ src/app/app.component.html
@@ -119,7 +119,7 @@
           </div>
       </div>
       <div class="col-xl-6">
-          <lib-fabricjs-library #canvas></lib-fabricjs-library>
+          <lib-angular-editor-fabric-js #canvas></lib-angular-editor-fabric-js>
       </div>
       <div class="col-xl-3">
           <div class="card">
