+++ src/app/app.component.spec.ts
@@ -16,16 +16,16 @@ describe('AppComponent', () => {
     expect(app).toBeTruthy();
   });
 
-  it(`should have as title 'fabricjs-editor'`, () => {
+  it(`should have as title 'angular-editor-fabric-js'`, () => {
     const fixture = TestBed.createComponent(AppComponent);
     const app = fixture.componentInstance;
-    expect(app.title).toEqual('fabricjs-editor');
+    expect(app.title).toEqual('angular-editor-fabric-js');
   });
 
   it('should render title', () => {
     const fixture = TestBed.createComponent(AppComponent);
     fixture.detectChanges();
     const compiled = fixture.nativeElement;
-    expect(compiled.querySelector('.content span').textContent).toContain('fabricjs-editor app is running!');
+    expect(compiled.querySelector('.content span').textContent).toContain('angular-editor-fabric-js app is running!');
   });
 });
