+++ src/app/app.module.ts
@@ -2,7 +2,7 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 
 import { AppComponent } from './app.component';
-import { FabricjsLibraryModule } from 'projects/fabricjs-library/src/public-api';
+import { FabricjsLibraryModule } from 'projects/angular-editor-fabric-js/src/public-api';
 import { FormsModule } from '@angular/forms';
 import { ColorPickerModule } from 'ngx-color-picker';
